requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
nltk==3.9.1
pyyaml==6.0.2
joblib==1.4.2
tqdm==4.66.5
rank-bm25==0.2.2
matplotlib==3.9.2
pyspellchecker==0.8.1
# CPU-only pins to avoid CUDA downloads
torch==2.2.2+cpu
torchvision==0.17.2+cpu
sentence-transformers==2.2.2
# Older hub/transformers to keep `cached_download` available
transformers==4.30.2
huggingface-hub==0.16.4
tokenizers==0.13.3
